refactor: 패키지명을 stringcalculator 로 변경
