+++ src/test/kotlin/stringcalculator/NumberCalculatorTest.kt
@@ -1,6 +1,5 @@
-package calculator
+package stringcalculator
 
-import org.assertj.core.api.Assertions
 import org.assertj.core.api.Assertions.*
 import org.junit.jupiter.api.Test
