+++ src/test/kotlin/stringcalculator/README.md
@@ -8,10 +8,11 @@
 ## 구현할 기능 목록  
 - [x] 숫자나 지정된 연산자 이외의 값이 포함되어 있을 경우 예외 발생  
 - [x] 입력은 `숫자 연산자 숫자 연산자 숫자...`와 같은 포맷이어야 함. 아닐 경우 예외 발생
-- [] 입력된 문자열을 적절히 파싱하여 연산을 수행할 수 있게 해야 함  
-  - 숫자, 연산자가 순서대로 입력될 것이기 때문에 연산자를 기준으로 split 하여 배열을 만들면 될듯함  
 - [x] 연산 기능 구현 
   - [x] 덧셈
   - [x] 뺄셈
   - [x] 곱셈 
   - [x] 나눗셈
+- [] 문자열 파싱 기능 구현 
+  - 숫자, 연산자가 순서대로 입력될 것이기 때문에 연산자를 기준으로 split 하여 배열을 만들면 될듯함
+- [] 숫자 연산 기능과 문자열 파싱 기능을 이용해 문자열 계산기 구현
