+++ src/test/kotlin/stringcalculator/StringCalculatorTest.kt
@@ -1,10 +1,7 @@
-package calculator
+package stringcalculator
 
-import org.assertj.core.api.Assertions
 import org.assertj.core.api.Assertions.*
-import org.junit.jupiter.api.Test
 import org.junit.jupiter.params.ParameterizedTest
-import org.junit.jupiter.params.provider.EmptySource
 import org.junit.jupiter.params.provider.NullAndEmptySource
 import org.junit.jupiter.params.provider.ValueSource
